see phab:T118683 for details.
For MediaWiki 1.27 and higher the third parameter needs to be declared as $user and not &$user. This is a known bug; see phab:T118683 for details. (If your hook function does not modify the User object, you can simply change the parameter to $user and it will work correctly across all MediaWiki versions.)